@@ -12,14 +12,16 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": ">=8.0",
+    "php": ">=8.3",
     "ext-intl": "*",
     "ext-tokenizer": "*",
+    "dibi/dibi": "^v5.0",
     "nette/di": "^3.1",
     "nette/neon": "^3.0",
     "nette/safe-stream": "^2.4|^3.0",
     "nette/utils": "^3.0|^4.0",
-    "psr/log": "^1.0"
+    "psr/log": "^1.0",
+    "symfony/console": "^7.2"
   },
   "require-dev": {
     "nette/tester": "^2.0",
