@@ -1,7 +1,7 @@
 <?php
 /**
  * @author    Miroslav Koula
- * @copyright Copyright (c) 2018 Miroslav Koula, https://elcheco.it
+ * @copyright Copyright (c) 2018 Miroslav Koula, https://koula.eu
  * @created  01/10/18 14:02
  */
 
@@ -9,46 +9,45 @@ declare(strict_types=1);
 
 namespace ElCheco\Translator;
 
-
 abstract class Dictionary implements DictionaryInterface
 {
+    /**
+     * @var array<string, string|array<int|string, string>>
+     */
+    private array $messages = [];
 
-	/**
-	 * @var array
-	 */
-	private $messages;
+    public function has(string $message): bool
+    {
+        $this->lazyLoad();
+        return array_key_exists($message, $this->messages);
+    }
 
+    /**
+     * @return string|array<int|string, string>
+     * @throws TranslatorException If message not found
+     */
+    public function get(string $message): string|array
+    {
+        $this->lazyLoad();
 
-	public function has(string $message): bool
-	{
-		$this->lazyLoad();
+        return $this->messages[$message] ?? throw new TranslatorException(
+            sprintf('Translation message "%s" not found.', $message)
+        );
+    }
 
-		return array_key_exists($message, $this->messages);
-	}
+    abstract protected function lazyLoad(): void;
 
+    protected function isReady(): bool
+    {
+        return !empty($this->messages);
+    }
 
-	public function get(string $message)
-	{
-		$this->lazyLoad();
-
-		return $this->messages[$message];
-	}
-
-
-	abstract protected function lazyLoad();
-
-
-	protected function isReady(): bool
-	{
-		return is_array($this->messages);
-	}
-
-
-	protected function setMessages(array $messages): DictionaryInterface
-	{
-		$this->messages = $messages;
-
-		return $this;
-	}
-
+    /**
+     * @param array<string, string|array<int|string, string>> $messages
+     */
+    protected function setMessages(array $messages): self
+    {
+        $this->messages = $messages;
+        return $this;
+    }
 }
