@@ -1,7 +1,7 @@
 <?php
 /**
  * @author    Miroslav Koula
- * @copyright Copyright (c) 2018 Miroslav Koula, https://elcheco.it
+ * @copyright Copyright (c) 2018 Miroslav Koula, https://koula.eu
  * @created  01/10/18 14:02
  */
 
